The coding is completed. Tests are the next iteration...

+++ src/main/java/mrbaxmypka/gmail/com/mapPointsTrimmer/entitiesDto/PlacemarkNodeDto.java
@@ -13,11 +13,41 @@ public class PlacemarkNodeDto {
 
     private Node placemarkNode;
 
+    /**
+     * The "coordinates" Node is a required node.
+     * {@literal Either <coordinates>-119.779550,33.829268,0</coordinates>
+     * or <coordinates>-122.000,37.002,127.00</coordinates> or <coordinates>-122.000,37.002</coordinates>}
+     * A single tuple consisting of floating point values for longitude, latitude, and altitude (in that order).
+     * Longitude and latitude values are in degrees, where
+     * longitude ≥ −180 and <= 180
+     * latitude ≥ −90 and ≤ 90
+     * altitude values (optional) are in meters above sea level.
+     */
     private double longitude;
 
+    /**
+     * The "coordinates" Node is a required node.
+     * {@literal Either <coordinates>-119.779550,33.829268,0</coordinates>
+     * or <coordinates>-122.000,37.002,127.00</coordinates> or <coordinates>-122.000,37.002</coordinates>}
+     * A single tuple consisting of floating point values for longitude, latitude, and altitude (in that order).
+     * Longitude and latitude values are in degrees, where
+     * longitude ≥ −180 and <= 180
+     * latitude ≥ −90 and ≤ 90
+     * altitude values (optional) are in meters above sea level.
+     */
     private double latitude;
 
-    private Integer altitude;
+    /**
+     * The "coordinates" Node is a required node.
+     * {@literal Either <coordinates>-119.779550,33.829268,0</coordinates>
+     * or <coordinates>-122.000,37.002,127.00</coordinates> or <coordinates>-122.000,37.002</coordinates>}
+     * A single tuple consisting of floating point values for longitude, latitude, and altitude (in that order).
+     * Longitude and latitude values are in degrees, where
+     * longitude ≥ −180 and <= 180
+     * latitude ≥ −90 and ≤ 90
+     * altitude values (optional) are in meters above sea level.
+     */
+    private double altitude;
 
     /**
      * The name of an icon for this Placemark (those may be from its Style or NormalStyle from StyleMap).
